+++ src/rumors.c
@@ -313,7 +313,7 @@ register struct monst *oracl;
 
 	Sprintf(qbuf,
 		"\"Wilt thou settle for a minor consultation?\" (%d %s)",
-		minor_cost, currency(minor_cost));
+		minor_cost, currency((long)minor_cost));
 	switch (ynq(qbuf)) {
 	    default:
 	    case 'q':
@@ -338,7 +338,7 @@ register struct monst *oracl;
 		    (oracle_cnt == 1 || oracle_flg < 0)) return 0;
 		Sprintf(qbuf,
 			"\"Then dost thou desire a major one?\" (%d %s)",
-			major_cost, currency(major_cost));
+			major_cost, currency((long)major_cost));
 		if (yn(qbuf) != 'y') return 0;
 #ifndef GOLDOBJ
 		u_pay = (u.ugold < (long)major_cost ? (int)u.ugold
